# diff: Derivative Performance