--- a/diff/fd/derivative.go
+++ b/diff/fd/derivative.go
@@ -6,8 +6,6 @@ package fd
 
 import (
 	"math"
-	"runtime"
-	"sync"
 )
 
 // Derivative estimates the derivative of the function f at the given location.
@@ -19,7 +17,7 @@ func Derivative(f func(float64) float64, x float64, settings *Settings) float64 
 	formula := Forward
 	step := formula.Step
 	var originValue float64
-	var originKnown, concurrent bool
+	var originKnown bool
 
 	// Use user settings if provided.
 	if settings != nil {
@@ -33,39 +31,16 @@ func Derivative(f func(float64) float64, x float64, settings *Settings) float64 
 		}
 		originKnown = settings.OriginKnown
 		originValue = settings.OriginValue
-		concurrent = settings.Concurrent
 	}
 
 	var deriv float64
-	if !concurrent || runtime.GOMAXPROCS(0) == 1 {
-		for _, pt := range formula.Stencil {
-			if originKnown && pt.Loc == 0 {
-				deriv += pt.Coeff * originValue
-				continue
-			}
-			deriv += pt.Coeff * f(x+step*pt.Loc)
-		}
-		return deriv / math.Pow(step, float64(formula.Derivative))
-	}
-
-	wg := &sync.WaitGroup{}
-	mux := &sync.Mutex{}
 	for _, pt := range formula.Stencil {
 		if originKnown && pt.Loc == 0 {
-			mux.Lock()
 			deriv += pt.Coeff * originValue
-			mux.Unlock()
 			continue
 		}
-		wg.Add(1)
-		go func(pt Point) {
-			defer wg.Done()
-			fofx := f(x + step*pt.Loc)
-			mux.Lock()
-			defer mux.Unlock()
-			deriv += pt.Coeff * fofx
-		}(pt)
+		deriv += pt.Coeff * f(x+step*pt.Loc)
 	}
-	wg.Wait()
+
 	return deriv / math.Pow(step, float64(formula.Derivative))
 }