--- a/diff/fd/derivative_test.go
+++ b/diff/fd/derivative_test.go
@@ -7,9 +7,12 @@ package fd
 import (
 	"math"
 	"testing"
+	"time"
 )
 
 var xSquared = func(x float64) float64 { return x * x }
+
+var xSquaredSlow = func(x float64) float64 { time.Sleep(500); return x * x }
 
 type testPoint struct {
 	f    func(float64) float64
@@ -141,5 +144,33 @@ func TestDerivativeDefault(t *testing.T) {
 		if math.Abs(test.ans-ans) > tol {
 			t.Errorf("Case %v: ans mismatch zero value: expected %v, found %v", i, test.ans, ans)
 		}
+	}
+}
+
+func BenchmarkDerivativeSerial(b *testing.B) {
+	// run the Derivative benchmark b.N times
+	for n := 0; n < b.N; n++ {
+		Derivative(xSquared, 10, &Settings{Concurrent: false})
+	}
+}
+
+func BenchmarkDerivativeConcurrent(b *testing.B) {
+	// run the Derivative benchmark b.N times
+	for n := 0; n < b.N; n++ {
+		Derivative(xSquared, 10, &Settings{Concurrent: true})
+	}
+}
+
+func BenchmarkDerivative2ndSerial(b *testing.B) {
+	// run the Derivative benchmark b.N times
+	for n := 0; n < b.N; n++ {
+		Derivative(xSquared, 10, &Settings{Formula: Forward2nd, Concurrent: false})
+	}
+}
+
+func BenchmarkDerivative2ndConcurrent(b *testing.B) {
+	// run the Derivative benchmark b.N times
+	for n := 0; n < b.N; n++ {
+		Derivative(xSquared, 10, &Settings{Formula: Forward2nd, Concurrent: true})
 	}
 }