@kortschak 2018-01-04
I have not touched this code, but it seems to me that a concurrent implementation that performs better is easily achievable with minimal effort here; instead of the goroutine-per-point approach used here which is obviously inefficient, `runtime.GOMAXPROCS(0)` goroutines should be started, each with a work unit of `n/runtime.GOMAXPROCS(0)` points, these can then either update `$GOMAXPROCS` elements of a slice to be summed after all workers return, or be reduced some other way.

BTW, `benchstat` should be given more than 1 replicate run (use `go test -run ^$ -bench <foo> -count 10`, e.g.).